% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{add.buckets}
\alias{add.buckets}
\title{Create buckets}
\usage{
add.buckets(track.bucket, hub.bucket)
}
\arguments{
\item{track.bucket}{Name to be given for track bucket on Amazon S3. (Must be all lower-case)}

\item{hub.bucket}{Name to be given for hub bucket on Amazon S3. (Must be all lower-case)}
}
\description{
This function creates a track and hub bucket on Amazon S3
}
\keyword{internal}
