% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dst.R
\name{calculate_dst}
\alias{calculate_dst}
\title{Calculates the error associated to all possible segments x\link{s:t} where 1<=s<=t<=N and N is the length of x}
\usage{
calculate_dst(x)
}
\arguments{
\item{x}{a series}
}
\value{
a ff matrix M such that M\link{s,t} corresponds to the error associated with segment x\link{s,t}
}
\description{
Calculates the error associated to all possible segments x\link{s:t} where 1<=s<=t<=N and N is the length of x
}
\examples{
x=c(rnorm(100,3,2),rnorm(100,10,2),rnorm(100,2,2),rnorm(100,-4,2),rnorm(100,6,2),rnorm(100,2,2),rnorm(100,8,2))
calculate_dst(x)
}
