% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dstat.R
\name{Dstat}
\alias{Dstat}
\title{Total sum of squares}
\usage{
Dstat(x, seg)
}
\arguments{
\item{x}{a series}

\item{x}{a series}
}
\value{
a total sum of squares
}
\description{
Total sum of squares
}
\examples{
x=c(rnorm(100,3,2),rnorm(100,10,2),rnorm(100,2,2),rnorm(100,-4,2),rnorm(100,6,2),rnorm(100,2,2),rnorm(100,8,2))
}
