% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{fill.hub.bucket}
\alias{fill.hub.bucket}
\title{Upload hub file to hub bucket}
\usage{
fill.hub.bucket(data.dir, hub.file, hub.bucket)
}
\arguments{
\item{data.dir}{Directory where the bigwig files}

\item{hub.file}{Hub file produced by hubR}

\item{hub.bucket}{Name of the hub bucket on Amazon S3.}
}
\description{
This function takes as input the data directory and file name for the hub file and uploads it to the hub bucket.
}
\keyword{internal}
