% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{fill.track.bucket}
\alias{fill.track.bucket}
\title{Populate track bucket with bigwigs}
\usage{
fill.track.bucket(data.dir, bigwigs, track.bucket)
}
\arguments{
\item{data.dir}{Directory where the bigwig files exist}

\item{bigwigs}{Names of the bigwig files}

\item{track.bucket}{Name of the track bucket on Amazon S3.}
}
\description{
This function takes as input the data directory and file names for the bigwig files and uploads them to the track bucket.
}
