% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_templates.R
\name{generate.composite.track.hub}
\alias{generate.composite.track.hub}
\title{Build composite track hub}
\usage{
generate.composite.track.hub(
  track.bucket,
  bw.df,
  anno.df,
  output.track.file,
  email
)
}
\arguments{
\item{track.bucket}{Amazon S3 bucket name that stores the .bw files}

\item{bw.df}{Data.frame holding all bigwig and track information supplied by user.}

\item{anno.df}{Data.frame holding all data level information supplied by user.}

\item{output.track.file}{Output track hub filename. Default: trackDB.txt}

\item{email}{Correspondence email}
}
\value{
A matrix of the infile
}
\description{
This function..
}
\keyword{internal}
