% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_signal.R
\name{model_signal}
\alias{model_signal}
\title{Calculates the fitted values of the segmentation model}
\usage{
model_signal(x, seg)
}
\arguments{
\item{x}{a series of quantitative values}

\item{seg}{a series of segment identifiers}
}
\value{
a series where fitted values correspond to mean per segment
}
\description{
Calculates the fitted values of the segmentation model
}
\examples{
x=c(rnorm(100,3,2),rnorm(100,10,2),rnorm(100,2,2),rnorm(100,-4,2),rnorm(100,6,2),rnorm(100,2,2),rnorm(100,8,2))
y=model_signal(x,seg=c(1,300,701))
plot(x)
lines(y, col="red")
}
