% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scheffe_test.R
\name{Scheffe_test}
\alias{Scheffe_test}
\title{This function tests the differences of consecutive segments}
\usage{
Scheffe_test(x, v_r, k = length(v_r - 1), alpha = 0.05)
}
\arguments{
\item{x}{a series}

\item{v_r}{the cuts locations in the series}
}
\value{
a test
}
\description{
This function tests the differences of consecutive segments
}
\examples{
x=c(rnorm(100,3,2),rnorm(100,10,2),rnorm(100,2,2),rnorm(100,-4,2),rnorm(100,6,2),rnorm(100,2,2),rnorm(100,8,2))
Scheffe_test(x,v_r=c(1,51,101,201,351,401,501,601,701))
}
