% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_signal.R
\name{segment_signal}
\alias{segment_signal}
\title{Produces a series of segment identifiers based on a vector listing the boundaries of the segments}
\usage{
segment_signal(seg)
}
\arguments{
\item{seg}{the boundaries of segments c(i1,i2,....,N) where i1, i2, ... correspond to first elements of each segment and N is the index of the last element}
}
\value{
a series corresponding to segment identifiers
}
\description{
Produces a series of segment identifiers based on a vector listing the boundaries of the segments
}
\examples{
segment_signal(c(1,40,68,100))
}
