% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{setup.aws}
\alias{setup.aws}
\title{Setup aws.s3}
\usage{
setup.aws(access.key, secret.key, region = "us-west-2")
}
\arguments{
\item{access.key}{Amazon S3 access key ID}

\item{secret.key}{Amazon S3 access key ID}

\item{region}{Amazon S3 region: default 'us-west-2'}
}
\description{
This function takes in the users AWS credentials and primes aws.s3.
}
