% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{setup.awscli}
\alias{setup.awscli}
\title{Setup awscli2}
\usage{
setup.awscli(
  awscli.path = "/allen/programs/celltypes/workgroups/hct/NelsonJ/Home/v2/2.4.9/bin"
)
}
\arguments{
\item{awscli.path}{directory where awscli2 was installed}
}
\description{
This function adds an awscli2 install to the users PATH
}
