% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_segmentation.R
\name{test_segmentation}
\alias{test_segmentation}
\title{Tests whether all consecutive segments are significantly different}
\usage{
test_segmentation(x, v_r, alpha = 0.05)
}
\arguments{
\item{x}{a series}

\item{v_r}{the cuts locations in the series}
}
\value{
a total sum of squares
}
\description{
Tests whether all consecutive segments are significantly different
}
\examples{
x=c(rnorm(100,3,2),rnorm(100,10,2),rnorm(100,2,2),rnorm(100,-4,2),rnorm(100,6,2),rnorm(100,2,2),rnorm(100,8,2))
test_segmentation(x, v_r=c(1,301,501,701))
}
