% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{update.bucket.permission}
\alias{update.bucket.permission}
\title{Update bucket permission}
\usage{
\method{update}{bucket.permission}(
  track.bucket,
  access.key = "AKIAXEERKKJCONE3EJVO",
  secret.key = "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"
)
}
\arguments{
\item{track.bucket}{Name of the track bucket on Amazon S3.}

\item{access.key}{Amazon S3 access key ID}

\item{secret.key}{Amazon S3 access key ID}
}
\description{
This function is intended to update the track bucket to be completly private
}
