% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub.R
\name{generate.multiwig.track.hub}
\alias{generate.multiwig.track.hub}
\title{Build multi-wig track hub}
\usage{
generate.multiwig.track.hub(
  hmac.encoded,
  track.bucket,
  bigwigs,
  pseudo.names,
  long.labels,
  colors,
  species,
  region,
  type,
  cluster,
  genome,
  output.track.file,
  email
)
}
\arguments{
\item{hmac.encoded}{Bigwig file signatures computed with \code{create.signatures}}

\item{track.bucket}{Amazon S3 bucket name that stores the .bw files}

\item{bigwigs}{Bigwig file names which exactly match those in 'track.bucket'}

\item{pseudo.names}{Short labels to give each track. If NULL the function attemps to gather from .bw file name: pseudo.names-*.bw}

\item{long.labels}{Detailed labels to give each track. If NULL the function attemps to gather from .bw file name: long.labels-*.bw}

\item{colors}{Colors (in R, G, B format) to give the tracks. If NULL colors will be auto-generated.}

\item{species}{Species information}

\item{region}{Brain region information}

\item{type}{Data type (ATAC, Multiome, etc.)}

\item{cluster}{Cluster label information}

\item{genome}{Genome information}

\item{output.track.file}{Output track hub filename. Default: trackDB.txt}

\item{email}{Correspondence email}
}
\value{
A matrix of the infile
}
\description{
This function ..
}
\keyword{internal}
