% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub.R
\name{hubR}
\alias{hubR}
\title{Builds a track hub for the UCSC genome browser}
\usage{
hubR(
  track.bucket,
  hub.bucket,
  bigwigs,
  pseudo.names = NULL,
  long.labels = NULL,
  colors = NULL,
  species,
  region,
  type,
  cluster,
  genome,
  data.dir,
  output.track.file = "trackDB.txt",
  email = ""
)
}
\arguments{
\item{track.bucket}{Amazon S3 bucket name that stores the .bw files}

\item{hub.bucket}{Amazon S3 bucket name that stores the track hub .txt file}

\item{bigwigs}{Bigwig file names which exactly match those in 'track.bucket'}

\item{pseudo.names}{Short labels to give each track. If NULL the function attemps to gather from .bw file name: pseudo.names-*.bw}

\item{long.labels}{Detailed labels to give each track. If NULL the function attemps to gather from .bw file name: long.labels-*.bw}

\item{colors}{Colors (in R, G, B format) to give the tracks. If NULL colors will be auto-generated.}

\item{species}{Species information}

\item{region}{Brain region information}

\item{type}{Data type (ATAC, Multiome, etc.)}

\item{cluster}{Cluster label information}

\item{genome}{Genome information, must be a valid UCSC genome browser genome.}

\item{data.dir}{Location of the bigwig files and where the trackDB.txt will be saved locally}

\item{output.track.file}{Output track hub filename. Default: trackDB.txt}

\item{email}{Correspondence email}
}
\value{
Hub URL
}
\description{
This builds the track hub and handles all uploading to AWS S3 buckets as named in track.bucket and hub.bucket.
Any changes to the bigwig file names or 'track.bucket' will require the regeneration of HMAC signatures and the hub file.
}
