% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{set.bucket.permissions}
\alias{set.bucket.permissions}
\title{Update bucket permission}
\usage{
set.bucket.permissions(track.bucket)
}
\arguments{
\item{track.bucket}{Name of the track bucket on Amazon S3.}
}
\description{
This function is intended to update the track bucket to be completly private
}
\keyword{internal}
